Introducing CoreEditorUtils
